Document the whole crate
This adds documentation for every publicly exposed item. It also adds
safety documentation for unsafe code.
