+++ src/lib.rs
@@ -1,4 +1,35 @@
 #![no_std]
+#![warn(
+    clippy::complexity,
+    clippy::correctness,
+    clippy::perf,
+    clippy::style,
+    clippy::missing_const_for_fn,
+    clippy::undocumented_unsafe_blocks,
+    missing_docs,
+    rust_2018_idioms
+)]
+
+//! Helper crate to write auto splitters for LiveSplit One's auto splitting runtime.
+//!
+//! # Example
+//!
+//! ```no_run
+//! # use asr::Process;
+//! #[no_mangle]
+//! pub extern "C" fn update() {
+//!     if let Some(process) = Process::attach("Notepad.exe") {
+//!         asr::print_message("Hello World!");
+//!         if let Ok(address) = process.get_module_address("Notepad.exe") {
+//!             if let Ok(value) = process.read::<u32>(address) {
+//!                 if value > 0 {
+//!                     asr::timer::start();
+//!                 }
+//!             }
+//!         }
+//!     }
+//! }
+//! ```
 
 mod runtime;
 #[cfg(feature = "signature")]
