+++ src/runtime/user_settings.rs
@@ -1,0 +1,20 @@
+//! This module allows you to add settings that the user can modify.
+
+use super::sys;
+
+/// Adds a new setting that the user can modify. This will return either the
+/// specified default value or the value that the user has set.
+#[inline]
+pub fn add_bool(key: &str, description: &str, default_value: bool) -> bool {
+    // SAFETY: We provide valid pointers and lengths to key and description.
+    // They are also guaranteed to be valid UTF-8 strings.
+    unsafe {
+        sys::user_settings_add_bool(
+            key.as_ptr(),
+            key.len(),
+            description.as_ptr(),
+            description.len(),
+            default_value,
+        )
+    }
+}
